requirements.txt
flask>=2.0.0
click>=8.0.0
pyyaml>=6.0
requests>=2.25.0
pytest>=7.0.0
pytest-cov>=3.0.0
